refactor: consolidate detectors into policies, add regex validation
- Merge detectors.py into policies.py to reduce file count
- Add ValueError on invalid custom regex patterns instead of silent skipping
- Update test to verify ValueError is raised with clear message
- Update all imports to use consolidated policies module
- All 85 tests pass, linting clean, type checking clean
